refactor: add app module folder
