+++ src/App.vue
@@ -1,11 +1,10 @@
-<script setup lang="ts"></script>
+<script setup lang="ts">
+import SubmissionPage from './app/submission/views/SubmissionPage.vue'
+</script>
 
 <template>
   <UApp>
-    <UContainer as="main" class="p-2 pb-12 max-w-xl">
-      <ContestExplanation />
-      <SubmissionForm />
-    </UContainer>
+    <SubmissionPage />
   </UApp>
 </template>
