+++ src/app/submission/components/CardChoice.vue
