+++ src/app/submission/components/ContestExplanation.vue
