+++ src/app/submission/views/SubmissionPage.vue
@@ -1,0 +1,11 @@
+<script setup lang="ts">
+import ContestExplanation from '../components/ContestExplanation.vue'
+import SubmissionForm from '../components/SubmissionForm.vue'
+</script>
+
+<template>
+  <UContainer as="main" class="p-2 pb-12 max-w-xl">
+    <ContestExplanation />
+    <SubmissionForm />
+  </UContainer>
+</template>
